Fix line counter maybe?

+++ source/script/scriptError.c
@@ -12,6 +12,6 @@ void printScriptError(u8 errLevel, char* message, ...) {
 	gfx_printf("\n\n[%s] ", (errLevel == SCRIPT_FATAL) ? "FATAL" : (errLevel == SCRIPT_PARSER_FATAL) ? "PARSE_FATAL" : "WARN");
 	gfx_vprintf(message, args);
 	if (errLevel < SCRIPT_WARN)
-		gfx_printf("\nError occured on or near line %d\n", scriptCurrentLine);
+		gfx_printf("\nError occured on or near line %d\n", (u32)scriptCurrentLine);
 	va_end(args);
 }
